Add /instance/<region>/<id>/ GET
the new endpoint returns handy information like the public ip and dns
address of the instance. I assume this can be used to update folks
means of ssh'ing into their machines, as aws does not hold onto the
same ip or dns for stopped instances. That is, unless you pay for an
elastic ip, which doesnt make a whole lot of sense here, as we're
trying to save money with workwork.

Also added some VERY basic tests, I will make a second pass at these
tomorrow with something more robust.